tuxs/pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools_scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "tuxs"
version = "0.1.0"
description = "图像生成工具集，包含 Kimi API 集成、表格渲染和 HTML 截图等功能"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "tuxiansheng"}
]
keywords = ["kimi", "table", "renderer", "screenshot", "comfyui"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]

dependencies = [
    "requests>=2.25.0",
    "python-dotenv>=0.19.0",
    "beautifulsoup4>=4.9.0",
    "selenium>=4.0.0",
    "Pillow>=8.0.0",
    "numpy>=1.19.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=3.0.0",
]

[project.urls]
Homepage = "https://github.com/yourusername/tuxs"
Repository = "https://github.com/yourusername/tuxs"

[tool.setuptools]
packages = ["tuxs", "tuxs.utils"]

[tool.setuptools.package-dir]
tuxs = "."

[tool.setuptools.package-data]
tuxs = ["utils/*.md"]
